requirements.txt
asgiref==3.3.4
autopep8==1.5.6
beautifulsoup4==4.9.3
certifi==2020.12.5
cfscrape==2.1.1
chardet==4.0.0
Django==3.2
django-environ==0.4.5
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
idna==2.10
lxml==4.6.3
psycopg2==2.8.6
pycodestyle==2.7.0
PyJWT==2.0.1
pytz==2021.1
requests==2.25.1
soupsieve==2.2.1
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.4
whitenoise==5.2.0
